Transition between two animations now works fine, even if the root bone is not centered at the origin on the XZ plane.
Though you aren't adviced to use this feature too much, if you need the move
the character's center of mess from the origin to an offseted point, you'd
rather create the animation yourself, than have the AnimatedMesh class create
the interpolation.
But of course, for cycles (like walking), creating the transition animation
and making it work right, can be a real pain, and in that case it's a lot
easier to use this feature (this is what it's for, after all).